Scale properly the altitude with the longitude
* min() function now uses/returns float instead of int

+++ include/utils.h
@@ -6,7 +6,7 @@
 /*   By: pnguyen- <pnguyen-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/25 17:52:44 by pnguyen-          #+#    #+#             */
-/*   Updated: 2024/02/13 15:25:25 by pnguyen-         ###   ########.fr       */
+/*   Updated: 2024/02/14 15:49:10 by pnguyen-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 
 void	put_pixel(t_img	*img, int x, int y, int color);
 void	init_trigo_table(t_trigo_table *table);
-int		min(int a, int b);
+float	min(float a, float b);
 int		my_abs(int nb);
 void	reset_map(t_fdf *fdf);
